+++ tests/LeanRoleTest.php
@@ -16,16 +16,32 @@ class LeanRoleTest extends PHPUnit_Framework_TestCase {
     }
 
     public function testInitializeRole() {
+        // It should not raise error
+        $role = new LeanRole();
+        $role = new LeanRole("_Role");
+
+        $role = new LeanRole(null, "id123");
+        $this->assertEquals("id123", $role->getObjectId());
+    }
+
+    public function testGetChildrenAsRelation() {
+        $role = new LeanRole();
+        $this->assertTrue($role->getUsers() instanceof LeanRelation);
+        $this->assertTrue($role->getRoles() instanceof LeanRelation);
+    }
+
+    public function testSaveRole() {
+        $role = new LeanRole();
+        $role->setName("admin");
+
         $acl = new LeanACL();
         $acl->setPublicWriteAccess(true); // so it can be destroyed
-        $role = new LeanRole("guest", $acl);
+        $role->setACL($acl);
+
         $role->save();
         $this->assertNotEmpty($role->getObjectId());
-
-        $childrenUsers = $role->getUsers();
-        $this->assertTrue($childrenUsers instanceof LeanRelation);
-        $childrenRoles = $role->getRoles();
-        $this->assertTrue($childrenRoles instanceof LeanRelation);
+        $this->assertTrue($role->getUsers() instanceof LeanRelation);
+        $this->assertTrue($role->getRoles() instanceof LeanRelation);
 
         $role->destroy();
     }